rust-api: Restructure snowcap stuff
Removes the re-export of the snowcap api because it's not stable
